test(mcp): reorganize batch-named test files by feature
Dissolve the fixes/lot4/lot5/lot6/review files into the feature files
they belong to, rename runner_fixes to test_runner.py, add
test_mcp_server_run_records.py, and replace 48 copies of the
KONFAI_MCP_WORKSPACES_ROOT setenv with a workspace_root fixture.
One near-duplicate describe_app test is merged into its twin.
