+++ konfai-mcp/tests/conftest.py
@@ -31,6 +31,14 @@ if str(TESTS_ROOT) not in sys.path:
 
 
 @pytest.fixture
+def workspace_root(tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> Path:
+    """Point KONFAI_MCP_WORKSPACES_ROOT at a per-test directory and return it."""
+    root = tmp_path / "workspaces"
+    monkeypatch.setenv("KONFAI_MCP_WORKSPACES_ROOT", str(root))
+    return root
+
+
+@pytest.fixture
 def load_mcp_server() -> Iterator[Callable[[], ModuleType]]:
     loaded_modules: list[ModuleType] = []
